[fix] move releasing into a script
The PR details code had a shell injection vulnerability and it was a bit
hard to manage and test in the Jenkinsfile itself.

This change moves the release code into a script where it can be more
easily tested. It also avoids any shell injection by only ever passing
the PR body in a pipe.